Changed ROB entry ID type to unsigned short. Created function for creating an ROB entry in ROB. Changed to short to insure valid ID generation and reuse.

+++ include/tomasulo/rob_entry.hpp
@@ -2,7 +2,7 @@
 #define ROB_ENTRY_HPP
 
 struct ROB_entry {
-     unsigned int id;
+     unsigned short id;
      std::uint16_t dest;
      std::uint16_t data;
      bool is_memory;
